Disable all crafting recipes if bucket or poisonivy is missing
Before this change, only a subset of recipes is disabled if poisonivy is missing, and bucket is a hard dependency.
For the player, recipes which use uncraftable items are usually useless.
Disabling all recipes if one of the two mods is missing allows me to turn bucket into an optional dependency.