Make the average fire rate more robust against dtime variation
If the timer is reset to 0 after each shot, any overshoot in the time difference is lost,
which can make the average fire rate dependent on the server step setting and noise.

To fix this for sufficiently small dtime values,
I set the timer to the overshoot modulo the fire rate.
This fix does not work if dtime is larger than the fire rate,
but in this case the game cannot adhere to the fire rate anyway.